+++ src/Platform.h
@@ -2,7 +2,6 @@
 
 #include <stdlib.h>
 #include <string.h>
-#include <stdint.h>
 
 #ifdef _WIN32
 #define WIN32_LEAN_AND_MEAN
@@ -63,4 +62,4 @@
 int initializePlatform(void);
 void cleanupPlatform(void);
 
-uint64_t PltGetMillis(void);
+unsigned long long PltGetMillis(void);
